Commit the more compact qrcode to our face image
This is a great saving, and shows that the other paths being squished in
the same manner should get us around another 1k saving.
-rw-rw-r-- 1 john   8542 2015-02-14 18:25 wip.svg
                     ↓↓
-rw-rw-r-- 1 john   7047 2015-02-25 14:02 wip.svg

Then of course some style changes are going to bloat it a little more,
as the vector/large stage for screen has yet to be made.